Rename the package from smoothing_plugins->online_signal_smoothing
