+++ moveit_core/filter_plugin_butterworth.xml
@@ -1,5 +1,5 @@
 <library path="moveit_butterworth_filter">
-  <class type="single_waypt_smoothing_plugins::ButterworthFilterPlugin" base_class_type="single_waypt_smoothing_plugins::SmoothingBaseClass">
+  <class type="online_signal_smoothing::ButterworthFilterPlugin" base_class_type="online_signal_smoothing::SmoothingBaseClass">
     <description>
     Simple Butterworth 1st order lowpass filter that only has one tuneable parameter and does not overshoot.
     </description>
