+++ moveit_core/filter_plugin_ruckig.xml
@@ -1,5 +1,5 @@
 <library path="moveit_ruckig_filter">
-  <class type="single_waypt_smoothing_plugins::RuckigFilterPlugin" base_class_type="single_waypt_smoothing_plugins::SmoothingBaseClass">
+  <class type="online_signal_smoothing::RuckigFilterPlugin" base_class_type="online_signal_smoothing::SmoothingBaseClass">
     <description>
     Jerk-limited smoothing with the Ruckig library.
     </description>
