+++ moveit_core/online_signal_smoothing/include/moveit/online_signal_smoothing/butterworth_filter.h
@@ -33,35 +33,64 @@
  *********************************************************************/
 
 /* Author: Andy Zelenak
-   Description: Jerk-limited smoothing with the Ruckig library.
+   Description: A first-order Butterworth low-pass filter. There is only one parameter to tune.
  */
 
 #pragma once
 
+#include <cstddef>
+
 #include <moveit/robot_model/robot_model.h>
-#include <moveit/single_waypt_smoothing_plugins/smoothing_base_class.h>
+#include <moveit/online_signal_smoothing/smoothing_base_class.h>
 
-#include <ruckig/ruckig.hpp>
-
-namespace single_waypt_smoothing_plugins
+namespace online_signal_smoothing
 {
-class RuckigFilterPlugin : public SmoothingBaseClass
+/**
+ * Class ButterworthFilter - Implementation of a signal filter to soften jerks.
+ * This is a first-order Butterworth low-pass filter. First-order was chosen for 2 reasons:
+ * - It doesn't overshoot
+ * - Computational efficiency
+ */
+class ButterworthFilter
 {
 public:
-  RuckigFilterPlugin(){};
+  /**
+   * Constructor.
+   * @param low_pass_filter_coeff Larger filter_coeff-> more smoothing of servo commands, but more lag.
+   * Rough plot, with cutoff frequency on the y-axis:
+   * https://www.wolframalpha.com/input/?i=plot+arccot(c)
+   */
+  ButterworthFilter(double low_pass_filter_coeff);
 
-  bool initialize(rclcpp::Node::SharedPtr node, const moveit::core::JointModelGroup& group, size_t num_dof,
-                  double timestep) override;
+  double filter(double new_measurement);
 
-  bool doSmoothing(std::vector<double>& /*unused*/, std::vector<double>& velocity_vector) override;
+  void reset(const double data);
+
+private:
+  static constexpr std::size_t FILTER_LENGTH = 2;
+  std::array<double, FILTER_LENGTH> previous_measurements_;
+  double previous_filtered_measurement_;
+  // Scale and feedback term are calculated from supplied filter coefficient
+  double scale_term_;
+  double feedback_term_;
+};
+
+// Plugin
+class ButterworthFilterPlugin : public SmoothingBaseClass
+{
+public:
+  ButterworthFilterPlugin(){};
+
+  bool initialize(rclcpp::Node::SharedPtr node, const moveit::core::JointModelGroup& group /*unused*/, size_t num_dof,
+                  double /*unused*/) override;
+
+  bool doSmoothing(std::vector<double>& position_vector, std::vector<double>& /*unused*/) override;
 
   bool reset(const std::vector<double>& joint_positions) override;
 
 private:
-  size_t num_dof_;
   rclcpp::Node::SharedPtr node_;
-  std::shared_ptr<ruckig::Ruckig<0, true /*print debug when failure*/>> ruckig_;
-  std::shared_ptr<ruckig::InputParameter<0>> ruckig_input_;
-  std::shared_ptr<ruckig::OutputParameter<0>> ruckig_output_;
+  std::vector<ButterworthFilter> position_filters_;
+  size_t num_dof_;
 };
-}  // namespace single_waypt_smoothing_plugins
+}  // namespace online_signal_smoothing
